Changes have been made to the Education Program completion and Badge earning processes. (#90)
* The relationships of some entities have been changed

* Database connections for Education Details, Sessions, Lessons, and Assignments have been established

* The email sending system with SMTP has been added, and password change and reset processes have been implemented.

* Edit - Corrections and improvements have been made.

* Changes have been made to the Education Program completion and Badge earning processes.

+++ Business/Dtos/Responses/AccountEducationProgramResponses/GetListAccountEducationProgramResponse.cs
@@ -5,5 +5,5 @@ public class GetListAccountEducationProgramResponse
     public Guid Id { get; set; }
     public string AccountName { get; set; }
     public string EducationProgramName { get; set; }
-    public int StatusPercent { get; set; }
+    public double StatusPercent { get; set; }
 }
